Adds support to firmware update via HTTP
- Adds the use of the LwM2M object resource '5/0/8' to consult
the protocol supported by the device, it decides the URI format
that will be sent to the device.
- Adds the use of the LwM2M object resource '/5/0/9' to consult
the firmware update's delivery method supported (PUSH method is not supported yet)
- Adds the possibility to configure the firmware storage directory via
environment variable (FIRMWARE_UPDATE_DATA_PATH)
- Adds a file server HTTP. the port can be configured via environment
variable (FIRMWARE_UPDATE_HTTP_PORT)
- Updates the README
- Updates the client demo to export the lwm2m object resource '5/0/8'
- Adds a new configuration to the client demo (-protocols) which gives
the possibility to the user configure the protocols supported by the
device
- Improves client demo log
- Adds a initial support to read LwM2M resources multidimentional

ATTENTION: HTTPS is not supported as protocol to transfer the firmware image.